Organize namespaces in Console project
