+++ src/CoolPlace.Console/Menus/IMenu.cs
@@ -1,11 +1,11 @@
-﻿namespace CoolPlace.Console
+﻿namespace CoolPlace.Console.Menus
 {
     public interface IMenu<T> where T : Enum
     {
         /// <summary>
         /// Selects the given menu option.
         /// </summary>
-        public void Choose(T option);
+        void Choose(T option);
         /// <summary>
         /// Prints the available menu options.
         /// </summary>
