+++ src/CoolPlace.Console/Program.cs
@@ -1,4 +1,6 @@
-﻿using Microsoft.Extensions.DependencyInjection;
+﻿using CoolPlace.Console.Menus;
+using CoolPlace.Console.Utilities;
+using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 
 namespace CoolPlace.Console
